Execute the problematic test on JVM only
