+++ tests/shared/src/test/scala/cats/effect/IOSpec.scala
@@ -23,7 +23,6 @@ import cats.laws.discipline.arbitrary._
 import cats.effect.implicits._
 import cats.effect.laws.AsyncTests
 import cats.effect.testkit.TestContext
-import cats.effect.std.Semaphore
 import cats.syntax.all._
 
 import org.scalacheck.Prop, Prop.forAll
@@ -36,8 +35,6 @@ import org.typelevel.discipline.specs2.mutable.Discipline
 
 import scala.concurrent.{ExecutionContext, TimeoutException}
 import scala.concurrent.duration._
-
-import java.util.concurrent.CancellationException
 
 class IOSpec extends IOPlatformSpecification with Discipline with ScalaCheck with BaseSpec {
   outer =>
@@ -1150,46 +1147,6 @@ class IOSpec extends IOPlatformSpecification with Discipline with ScalaCheck wit
         }
 
         test.flatMap(_ => IO(canceled)) must completeAs(true)
-      }
-
-      "cancel all inner effects when canceled" in ticked { implicit ticker =>
-        val deadlock = for {
-          gate1 <- Semaphore[IO](2)
-          _ <- gate1.acquireN(2)
-
-          gate2 <- Semaphore[IO](2)
-          _ <- gate2.acquireN(2)
-
-          io = IO {
-            // these finalizers never return, so this test is intentionally designed to hang
-            // they flip their gates first though; this is just testing that both run in parallel
-            val a = (gate1.release *> IO.never) onCancel {
-              gate2.release *> IO.never
-            }
-
-            val b = (gate1.release *> IO.never) onCancel {
-              gate2.release *> IO.never
-            }
-
-            a.unsafeRunAndForget()
-            b.unsafeRunAndForget()
-          }
-
-          _ <- io.flatMap(_ => gate1.acquireN(2)).start
-          _ <- gate2.acquireN(2) // if both are not run in parallel, then this will hang
-        } yield ()
-
-        val test = for {
-          t <- IO(deadlock.unsafeToFutureCancelable())
-          (f, ct) = t
-          _ <- IO.fromFuture(IO(ct()))
-          _ <- IO.blocking(scala.concurrent.Await.result(f, Duration.Inf))
-        } yield ()
-
-        test.attempt.map {
-          case Left(t) => t.isInstanceOf[CancellationException]
-          case Right(_) => false
-        } must completeAs(true)
       }
 
     }
